+++ retrofit/WebServices/src/main/res/values/ids.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <item name="images_loader" type="id">1</item>
+</resources>
